refactor CardProcessor to DocProcessor
