+++ ingest/ingest_pkg/elasticsearch_utils.py
@@ -1,7 +1,8 @@
 import time
-from typing import Type, Optional
+from typing import Type, Optional, Dict
 
 from dotenv import dotenv_values
+from elasticsearch.helpers import bulk
 from elasticsearch_dsl import Index, Document
 from elasticsearch_dsl.connections import connections
 
@@ -31,3 +32,38 @@ class IndexMover:
         self.new_es_index.put_alias(name=self.alias)
         if self.old_es_index:
             self.old_es_index.delete()
+
+
+class DocProcessor:
+    def __init__(self, batch_size=500, dryrun: bool = False):
+        self.batch_size = batch_size
+        self.dryrun = dryrun
+        self.client = connections.get_connection()
+        self._movers: Dict[str, IndexMover] = {}
+        self._items: list[dict] = []
+
+    def register_index(self, doctype: Type[Document]):
+        """Mark an index as being updated, creating an IndexMover instance."""
+        key = doctype.Index.name
+        if key in self._movers:
+            return
+        if not self.dryrun:
+            self._movers[key] = IndexMover(doctype)
+
+    def add(self, doc: Document):
+        d = doc.to_dict(include_meta=True)
+        if not self.dryrun:
+            d["_index"] = self._movers[d["_index"]].new_name
+        self._items.append(d)
+        if len(self._items) > self.batch_size:
+            self.flush()
+
+    def flush(self):
+        if len(self._items) > 0 and not self.dryrun:
+            bulk(self.client, self._items)
+        self._items = []
+
+    def finalize(self):
+        self.flush()
+        for mover in self._movers.values():
+            mover.move_alias_to_new()
